refactor(dashboard-rework): Moved the rarity query down
