+++ apps/website/src/app/events/[code]/dashboard/(overview)/page.tsx
@@ -6,17 +6,17 @@ import {
   getTotalChests,
 } from "@/actions/chest";
 import { getEventByCode } from "@/actions/event";
-import { getChestCountPerRarity } from "@/actions/rarity";
 import { getChestCountPerReward } from "@/actions/reward";
 import ChestCountAccountChart from "@/components/chest-count-account-chart";
 import ChestCountCategoryChart from "@/components/chest-count-category-chart";
-import ChestCountRarityChart from "@/components/chest-count-rarity-chart";
 import ChestCountRewardChart from "@/components/chest-count-reward-chart";
 import LatestChestCard from "@/components/latest-chest-card";
 import PeakOpeningHourCard from "@/components/peak-opening-hour-card";
 import TotalChestCard from "@/components/total-chest-card";
 import EventInformationItem from "@/components/event-information-item";
 import EventNotFound from "@/components/event-not-found";
+import RarityCard from "@/components/rarity-card";
+import { FilterConfig } from "@/types/common";
 
 export default async function Page({
   params,
@@ -31,6 +31,10 @@ export default async function Page({
     return <EventNotFound />;
   }
 
+  const filters = {
+    eventId: event.id,
+  } satisfies FilterConfig;
+
   const chestCount = await getTotalChests({
     eventId: event.id,
   });
@@ -39,9 +43,6 @@ export default async function Page({
     eventId: event.id,
   });
   const peakOpeningHourData = await getPeakOpeningHourData({
-    eventId: event.id,
-  });
-  const rarities = await getChestCountPerRarity({
     eventId: event.id,
   });
   const categories = await getChestCountPerCategory({
@@ -69,7 +70,7 @@ export default async function Page({
         />
       </div>
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
-        <ChestCountRarityChart rarities={rarities} />
+        <RarityCard filters={filters} />
         <ChestCountCategoryChart categories={categories} />
       </div>
       <ChestCountRewardChart rewards={rewards} />
