+++ apps/website/src/app/page.tsx
@@ -6,28 +6,29 @@ import {
   getPeakOpeningHourData,
 } from "@/actions/chest";
 import { getChestCountPerEvent, getPossibleChestCount } from "@/actions/event";
-import { getChestCountPerRarity } from "@/actions/rarity";
 import { getChestCountPerReward } from "@/actions/reward";
 import ChestCountAccountChart from "@/components/chest-count-account-chart";
 import ChestCountCategoryChart from "@/components/chest-count-category-chart";
 import ChestCountEventTable from "@/components/chest-count-event-table";
-import ChestCountRarityChart from "@/components/chest-count-rarity-chart";
 import ChestCountRewardChart from "@/components/chest-count-reward-chart";
 import LatestChestCard from "@/components/latest-chest-card";
 import PeakOpeningHourCard from "@/components/peak-opening-hour-card";
+import RarityCard from "@/components/rarity-card";
 import TotalChestCard from "@/components/total-chest-card";
+import { FilterConfig } from "@/types/common";
 
 export default async function Dashboard() {
   const chestCount = await getTotalChests();
   const accountCount = await getTotalAccounts();
   const possibleChestCount = await getPossibleChestCount();
   const accounts = await getChestCountPerAccount();
-  const rarities = await getChestCountPerRarity();
   const chest = await getLatestChest();
   const categories = await getChestCountPerCategory();
   const events = await getChestCountPerEvent();
   const rewards = await getChestCountPerReward();
   const peakOpeningHourData = await getPeakOpeningHourData();
+
+  const filters = {} satisfies FilterConfig;
 
   return (
     <div className="grid grid-cols-1 gap-4">
@@ -43,7 +44,7 @@ export default async function Dashboard() {
         />
       </div>
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
-        <ChestCountRarityChart rarities={rarities} />
+        <RarityCard filters={filters} />
         <ChestCountCategoryChart categories={categories} />
       </div>
       <ChestCountRewardChart rewards={rewards} />
