+++ apps/website/src/types/common.d.ts
@@ -1,0 +1,4 @@
+export type FilterConfig = {
+  accountId?: number;
+  eventId?: number;
+};
